[Backend] Split target helpers by backend
