+++ testing/python/amd/test_tilelang_ds_read_tr.py
@@ -15,7 +15,8 @@ import torch
 import tilelang
 import tilelang.language as T
 import tilelang.testing
-from tilelang.backend.target import target_is_gfx950, determine_target
+from tilelang.backend.target import determine_target
+from tilelang.rocm.target import target_is_gfx950
 
 
 def requires_gfx950():
