+++ testing/python/amd/test_tilelang_gemm_mfma_intrinsic.py
@@ -8,7 +8,8 @@ from tilelang.rocm.intrinsics.mfma_macro_generator import (
     MatrixCoreIntrinEmitter,
 )
 from tilelang.transform import simplify_prim_func
-from tilelang.backend.target import determine_fp8_type, determine_target
+from tilelang.backend.target import determine_target
+from tilelang.language.fp8 import determine_fp8_type
 
 tilelang.testing.set_random_seed(0)
