+++ testing/python/language/test_tilelang_language_get_warp_info.py
@@ -3,7 +3,7 @@ from typing import Optional
 import tilelang.language as T
 import tilelang.testing
 import torch
-from tilelang.backend.target import check_hip_availability
+from tilelang.rocm.target import check_hip_availability
 
 _IS_HIP_AVAILABLE = check_hip_availability()
 _DEFAULT_WARPS_PER_GROUP = 4
