+++ tilelang/cuda/op/gemm/__init__.py
@@ -8,7 +8,7 @@ from .gemm_mma_sm70 import GemmMMASm70
 from .gemm_mma_sm75 import GemmMMASm75
 from .gemm_tcgen05 import GEMM_INST_TCGEN05, GemmTCGEN5
 from .gemm_wgmma import GEMM_INST_WGMMA, GemmWGMMA
-from tilelang.backend.target import target_is_cuda, target_is_turing, target_is_volta
+from tilelang.cuda.target import target_is_cuda, target_is_turing, target_is_volta
 
 
 def _match_mma(target) -> bool:
