+++ tilelang/jit/adapter/cutedsl/adapter.py
@@ -14,7 +14,7 @@ from tilelang.jit.adapter.wrapper import TLPyWrapper
 from tilelang.jit.adapter.cutedsl.checks import check_cutedsl_available
 from tilelang.jit.adapter.cutedsl.libgen import CuTeDSLLibraryGenerator
 from tilelang.utils.language import retrieve_func_from_module
-from tilelang.backend.target import determine_target, normalize_cutedsl_target
+from tilelang.backend.target import determine_target
 from tilelang.jit.adapter.base import BaseKernelAdapter, CachedTextSource
 
 logger = logging.getLogger(__name__)
@@ -152,6 +152,8 @@ class CuTeDSLKernelAdapter(BaseKernelAdapter):
             adapter.param_shapes.append(native_shape)
 
         adapter.dynamic_symbolic_map, adapter.dynamic_symbolic_order = adapter._process_dynamic_symbolic()
+
+        from tilelang.cuda.target import normalize_cutedsl_target
 
         adapter.target = normalize_cutedsl_target(target) or Target(determine_target(target))
         adapter.verbose = verbose
