+++ tilelang/metal/target.py
@@ -2,14 +2,34 @@ from __future__ import annotations
 
 from platform import mac_ver
 
-from tilelang.backend.target import register_target_detector
+from tvm.target import Target
+
+from tilelang.backend.target import register_supported_target, register_target_detector
+
+
+def _target_ffi_api():
+    from tilelang import _ffi_api
+
+    return _ffi_api
+
+
+def check_metal_availability() -> bool:
+    mac_release, _, arch = mac_ver()
+    if not mac_release:
+        return False
+    # todo: check torch version?
+    return arch == "arm64"
 
 
 def _detect_metal_target() -> str | None:
-    mac_release, _, arch = mac_ver()
-    if mac_release and arch == "arm64":
+    if check_metal_availability():
         return "metal"
     return None
 
 
+def target_is_metal(target: Target) -> bool:
+    return _target_ffi_api().TargetIsMetal(target)
+
+
 register_target_detector("metal", _detect_metal_target, override=True)
+register_supported_target("metal", "Apple Metal target for arm64 Macs.", override=True)
